requirements.txt
Django==5.1.4
django-celery==3.1.17
django-crispy-forms==2.3
crispy-bootstrap4==2024.10
pillow==11.1.0
pywhatkit==5.4
django-cors-headers==4.6.0
django-environ==0.11.2
mysqlclient==2.2.6
django-simple-history==3.7.0

